refactor auth-related parts of config manager into auth manager, add error handling for `configure` calls to `config_changed`
